requirements.txt
streamlit
Pillow
colorthief
